Add push for safe-lists
Also fix a wrong signature for push for the unsafe list version.

+++ src/list/safe_u32.rs
@@ -1,5 +1,5 @@
 // TODO: Replace this when we have dynamical allocation
-pub const SAFE_LIST_ELEMENT_CAPACITY: usize = 1000;
+pub const SAFE_LIST_ELEMENT_CAPACITY: u32 = 1000;
 
 pub mod get;
 pub mod length;
